+++ src/briefy/leica/models/job/location.py
@@ -13,13 +13,46 @@ import sqlalchemy_utils as sautils
 class LocationContactInfoMixin(NameMixin):
     """A mixin to manage contact information for the order location."""
 
-    email = sa.Column(sautils.types.EmailType(), nullable=True, unique=False)
+    email = sa.Column(
+        sautils.types.EmailType(),
+        nullable=True,
+        unique=False,
+        info={
+            'colanderalchemy': {
+                'title': 'Email',
+                'missing': colander.drop,
+            }
+        }
+    )
     """Email of the contact person."""
 
-    mobile = sa.Column(sautils.types.PhoneNumberType(), nullable=True, unique=False)
+    mobile = sa.Column(
+        sautils.types.PhoneNumberType(),
+        nullable=True,
+        unique=False,
+        info={
+            'colanderalchemy': {
+                'title': 'Mobile Phone',
+                'typ': colander.String,
+                'missing': colander.drop,
+            }
+        }
+
+    )
     """Mobile phone number of the contact person."""
 
-    additional_phone = sa.Column(sautils.types.PhoneNumberType(), nullable=True, unique=False)
+    additional_phone = sa.Column(
+        sautils.types.PhoneNumberType(),
+        nullable=True,
+        unique=False,
+        info={
+            'colanderalchemy': {
+                'title': 'Additional Phone',
+                'typ': colander.String,
+                'missing': colander.drop,
+            }
+        }
+    )
     """Additional phone number of the contact person."""
 
 
@@ -52,3 +85,9 @@ class OrderLocation(LocationContactInfoMixin, AddressMixin,
 
     Reference to the Order this location is attached to.
     """
+
+    def to_dict(self):
+        """Custom to_dict method."""
+        data = super().to_dict()
+        data['coordinates'] = self.coordinates
+        return data