+++ HISTORY.rst
@@ -80,6 +80,7 @@ History
      * Add new field for Professional: accept_travel (boolean) (rudaporto).
      * Remove all binary=false from UUID fields (rudaporto).
      * Adjust users sync to update Knack Profile.briefy_id if not equal to same user.briefy_id in Rolleiflex (rudaporto).
+     * New descriptor to help set and get from unary relationships like Order.location (rudaporto).
 
 1.1.0 (2016-10-04)
 ------------------